fix: 修复 Windows 下类型收窄编译错误
MSVC 不允许将 qsizetype 通过列表初始化隐式收窄到 int。

为 SVG 清理替换区间增加显式 static_cast<int>，保持现有替换逻辑不变。

修改文件：

- src/services/ZhipuPPTAgentService.cpp

+++ src/services/ZhipuPPTAgentService.cpp
@@ -800,7 +800,8 @@ QString ZhipuPPTAgentService::sanitizeSvg(const QString &svgCode) const
         cleaned.remove(unsupportedPropRe);
 
         if (cleaned != styleBlock) {
-            replacements.append({match.capturedStart(), {match.capturedLength(), cleaned}});
+            replacements.append({static_cast<int>(match.capturedStart()),
+                                 {static_cast<int>(match.capturedLength()), cleaned}});
         }
     }
     // 从后向前替换
@@ -826,7 +827,8 @@ QString ZhipuPPTAgentService::sanitizeSvg(const QString &svgCode) const
             QRegularExpression::CaseInsensitiveOption);
         cleaned.remove(cssFilterRe);
         if (cleaned != styleVal) {
-            replacements.append({match.capturedStart(), {match.capturedLength(),
+            replacements.append({static_cast<int>(match.capturedStart()),
+                                 {static_cast<int>(match.capturedLength()),
                                  QStringLiteral("style=\"%1\"").arg(cleaned)}});
         }
     }
